% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyze}
\alias{analyze}
\title{Call QUADrATiC with a query gene signature}
\usage{
analyze(query.gene.sig, analysis.set = "Drug_Name", num.rand.sigs = 2000,
  endpoint = getOption("quadratic.endpoint", default =
  "http://127.0.0.1:8090"))
}
\arguments{
\item{query.gene.sig}{A data frame containing one column of probe IDs and a second with values of 1 where up-regulated and -1 where down-regulated.}

\item{analysis.set}{Define how reference profiles are grouped into reference sets in QUADrATiC. Valid values are Drug_Name, Drug_Name+Cell_Line, Drug_Name+Cell_Line+Time, Drug_Name+Conc+Cell_Line and Drug_Name+Conc+Cell_Line+Time. Default is Drug_Name.}

\item{num.rand.sigs}{The number of random signatures generated to estimate p-value. Default is 2000.}

\item{endpoint}{The URL for QUADrATiC. Default is http://localhost:8090. This can also be set globally by setting the option 'quadratic.endpoint'.}
}
\description{
Call QUADrATiC with a query gene signature
}
\examples{
estrogen.signature <- read.table('data/Estrogen.tsv', stringsAsFactors = F)
analyze(estrogen.signature)

}
